+++ src/elements/author/src/stories.tsx
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import * as React from 'react'
 import { jsx } from '@emotion/core'
-import { date, text } from '@storybook/addon-knobs'
+import { text } from '@storybook/addon-knobs'
 
 import AllThemes from '../../../utils/all-themes'
 
@@ -12,44 +12,25 @@ export default {
 }
 
 export const Example = () => {
-  const name = text('Author name', 'Salman')
-  const role = text('Author role', 'Job title')
-  const editorName = text('Editor name', 'Sarah Guershon')
-  const editorRole = text('Editor role', 'Job title')
-  const authorUrl = text('Author role', 'https://www.jonathanfielding.com/')
-  const editorUrl = text('Author role', 'https://www.jonathanfielding.com/')
+  const title = text('Title', 'Salman')
+  const subtitle = text('Subtitle', 'Job title')
+  const titleUrl = text('Title url', 'https://www.jonathanfielding.com/')
+  const subtitleUrl = text('Subtitle url', 'https://www.jonathanfielding.com/')
   const image = (
     <img src="https://placekitten.com/200/200?image=9" alt="Salman" />
   )
-  const updatedDate = date('Article date', new Date(2020, 2, 15))
 
   return (
     <div>
-      <h3>Without editor</h3>
+      <h3>Without subtitle</h3>
+      <Author title={title} authorImage={image} titleUrl={titleUrl} />
+      <h3>With subtitle</h3>
       <Author
-        name={name}
-        role={role}
+        title={title}
+        subtitle={subtitle}
         authorImage={image}
-        authorUrl={authorUrl}
-        date={updatedDate}
-      />
-      <h3>Without date</h3>
-      <Author
-        name={name}
-        role={role}
-        authorImage={image}
-        authorUrl={authorUrl}
-      />
-      <h3>With editor</h3>
-      <Author
-        name={name}
-        role={role}
-        editorName={editorName}
-        editorRole={editorRole}
-        authorImage={image}
-        authorUrl={authorUrl}
-        editorUrl={editorUrl}
-        date={updatedDate}
+        titleUrl={titleUrl}
+        subtitleUrl={subtitleUrl}
       />
     </div>
   )
@@ -61,10 +42,48 @@ Example.story = {
   }
 }
 
+export const CondensedExample = () => {
+  const title = text('Title', 'Salman')
+  const subtitle = text('Subtitle', 'Job title')
+  const titleUrl = text('Title url', 'https://www.jonathanfielding.com/')
+  const subtitleUrl = text('Subtitle url', 'https://www.jonathanfielding.com/')
+  const image = (
+    <img src="https://placekitten.com/200/200?image=9" alt="Salman" />
+  )
+
+  return (
+    <div>
+      <h3>Without subtitle</h3>
+      <Author
+        variant="condensed"
+        title={title}
+        authorImage={image}
+        titleUrl={titleUrl}
+      />
+      <h3>With subtitle</h3>
+      <Author
+        variant="condensed"
+        title={title}
+        subtitle={subtitle}
+        authorImage={image}
+        titleUrl={titleUrl}
+        subtitleUrl={subtitleUrl}
+      />
+    </div>
+  )
+}
+
+CondensedExample.story = {
+  parameters: {
+    percy: { skip: true }
+  }
+}
+
 export const AutomatedTests = () => {
   return (
-    <AllThemes themes={['uswitch', 'money']}>
+    <AllThemes>
       <Example />
+      <CondensedExample />
     </AllThemes>
   )
 }