+++ src/elements/breadcrumbs/src/index.tsx
@@ -6,7 +6,9 @@ import get from '@uswitch/trustyle-utils.get'
 import { Icon } from '@uswitch/trustyle.icon'
 
 const lookup = (variant: string) =>
-  variant === 'base' ? 'breadcrumbs2.base' : `breadcrumbs2.variants.${variant}`
+  variant === 'base'
+    ? 'elements.breadcrumbs.base'
+    : `elements.breadcrumbs.variants.${variant}`
 
 interface HomeIconProps extends React.HTMLAttributes<HTMLDivElement> {
   variant?: 'base' | 'light'
@@ -68,6 +70,7 @@ const MobileBreadcrumbs: React.FC<Props> = ({
         display: 'inline-block',
         color: 'inherit',
         textDecoration: 'none',
+        borderBottom: 0,
         fontSize: 'xxs',
         variant: `${lookup(variant)}.mobileLink`
       }}
@@ -103,6 +106,7 @@ const DesktopBreadcrumbs: React.FC<Props> = ({
     color: 'inherit',
     stroke: 'inherit',
     textDecoration: 'none',
+    borderBottom: 0,
     variant: `${lookup(variant)}.a`,
     ':visited': {
       color: 'inherit'