+++ src/elements/hero/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.1.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero@3.0.0...@uswitch/trustyle.hero@3.1.0) (2020-07-20)
+
+
+### Features
+
+* add form and rich text block panel styling to theme migration ([10c9e3f](https://github.com/uswitch/trustyle/commit/10c9e3f))
+
+
+
+
+
+# [3.0.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero@2.6.0...@uswitch/trustyle.hero@3.0.0) (2020-07-17)
+
+**Note:** Version bump only for package @uswitch/trustyle.hero
+
+
+
+
+
+# [2.6.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero@2.2.0...@uswitch/trustyle.hero@2.6.0) (2020-07-15)
+
+
+### Bug Fixes
+
+* hero breadcrumb styling for Money ([bf296a9](https://github.com/uswitch/trustyle/commit/bf296a9))
+
+
+### Features
+
+* update hero padding bottom ([69a339b](https://github.com/uswitch/trustyle/commit/69a339b))
+
+
+
+
+
 # [2.5.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.hero@2.2.0...@uswitch/trustyle.hero@2.5.0) (2020-07-13)